requirement.txt
absl-py==0.7.1
astor==0.7.1
atari-py==0.1.7
tensorflow-gpu==1.12.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
Cython==0.29.7
dill==0.2.9
future==0.17.1
gast==0.2.2
glfw==1.7.1
grpcio==1.20.1
gym==0.12.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lockfile==0.12.2
Markdown==3.1
mock==2.0.0
mpi4py==3.0.1
numpy==1.16.3
opencv-python==4.1.0.25
pbr==5.2.0
Pillow==6.0.0
pkg-resources==0.0.0
progressbar2==3.39.3
protobuf==3.7.1
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
python-utils==2.3.0
pyzmq==18.0.1
requests==2.21.0
scipy==1.2.1
six==1.12.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.2
Werkzeug==0.15.2
zmq==0.0.0
